requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
Django==6.0.1
dj-database-url==1.2.0
django-crispy-forms==2.5
gunicorn==21.2.0
Pillow==12.1.0
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.5
tinycss2==1.5.1
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.4.0
